cosme: cut functions into a module
